requirements.txt
fastapi
uvicorn[standard]
chromadb
pypdf
python-docx
docx2txt
pytesseract
pillow
pdf2image
requests
PyPDF2
